+++ src/utilities/fsUtils.ts
@@ -1,8 +1,24 @@
 import * as vscode from 'vscode';
+import { Uri, Webview } from 'vscode';
 import { promisify } from 'util';
 import { exec, spawn } from 'child_process';
 
 const execAsync = promisify(exec);
+/**
+ *
+ * A helper function which will get the webview URI of a given file or resource.
+ *
+ * @remarks This URI can be used within a webview's HTML as a link to the
+ * given file/resource.
+ *
+ * @param webview A reference to the extension webview
+ * @param extensionUri The URI of the directory containing the extension
+ * @param pathList An array of strings representing the path to a file/resource
+ * @returns A URI pointing to the file/resource
+ */
+export function getUri(webview: Webview, extensionUri: Uri, pathList: string[]) {
+    return webview.asWebviewUri(Uri.joinPath(extensionUri, ...pathList));
+}
 
 /**
  * Reads a text file from the given URI.
@@ -43,7 +59,8 @@ export async function getPythonPath(): Promise<string | undefined> {
     const pythonExtension = vscode.extensions.getExtension('ms-python.python');
     if (pythonExtension) {
         await pythonExtension.activate();
-        return pythonExtension.exports.settings.getExecutionDetails().execCommand.join(' ');
+        const details = await pythonExtension.exports.settings.getExecutionDetails();
+        return details.execCommand.join(' ');
     }
     // otherwise fallback to the python path from the python config
     const pythonConfig = vscode.workspace.getConfiguration('python');